+++ src/view/app.py
@@ -22,6 +22,7 @@ from typing import (Any, Callable, Coroutine, Generic, TextIO, TypeVar,
 from urllib.parse import urlencode
 
 import ujson
+import uvicorn
 from rich import print
 from rich.traceback import install
 
@@ -32,9 +33,9 @@ from ._loader import finalize, load_fs, load_simple
 from ._logging import (Internal, Service, UvicornHijack, enter_server,
                        exit_server, format_warnings)
 from ._parsers import supply_parsers
-from ._util import attempt_import, make_hint
+from ._util import make_hint
 from .config import Config, load_config
-from .exceptions import MissingLibraryError, ViewError
+from .exceptions import ViewError
 from .routing import Route, RouteOrCallable, V, _NoDefault, _NoDefaultType
 from .routing import body as body_impl
 from .routing import delete, get, options, patch, post, put
@@ -429,20 +430,18 @@ class App(ViewApp):
         uvloop_enabled = False
 
         if self.config.app.uvloop is True:
-            uvloop = attempt_import("uvloop")
+            uvloop = importlib.import_module("uvloop")
             uvloop.install()
             uvloop_enabled = True
         elif self.config.app.uvloop == "decide":
-            with suppress(MissingLibraryError):
-                uvloop = attempt_import("uvloop")
+            with suppress(ModuleNotFoundError):
+                uvloop = importlib.import_module("uvloop")
                 uvloop.install()
                 uvloop_enabled = True
 
         start = start_target or asyncio.run
 
         if server == "uvicorn":
-            uvicorn = attempt_import("uvicorn")
-
             config = uvicorn.Config(
                 self._app,
                 port=self.config.server.port,
@@ -459,7 +458,7 @@ class App(ViewApp):
             return start(self._spawn(server.serve()))
 
         elif server == "hypercorn":
-            hypercorn = attempt_import("hypercorn")
+            raise NotImplementedError
             conf = hypercorn.Config()
             conf.loglevel = "debug" if self.config.dev else "info"
             conf.bind = [