+++ src/view/databases.py
@@ -1,106 +1,25 @@
 from __future__ import annotations
 
+import asyncio
+import sqlite3
 from abc import ABC, abstractmethod
-from typing import (
-    Any,
-    Callable,
-    ClassVar,
-    Generic,
-    TypeVar,
-    cast,
-    overload,
-    Dict,
-)
+from enum import Enum
+from typing import Any, ClassVar, TypeVar, Union, get_type_hints
 
-from typing_extensions import ParamSpec, dataclass_transform
+import mysql.connector
+import psycopg2
+import pymongo
+from typing_extensions import Annotated, Self, dataclass_transform
 
-from ._util import attempt_import, make_hint
-from .exceptions import MistakeError
+from .routing import BodyParam
+from .typing import ViewBody
 
-P = ParamSpec("P")
-A = TypeVar("A")
-
-
-class ViewModel:
-    _conn: ClassVar[_Connection | None]
-
-
-T = TypeVar("T", bound=ViewModel)
-
-Where = Dict[str, Any]
-Value = Dict[str, Any]
-
-
-class Model(Generic[P, T]):
-    def __init__(
-        self,
-        ob: Callable[P, T],
-        table: str,
-        *,
-        conn: _Connection | None = None,
-    ) -> None:
-        self.obj = ob
-        self._conn: _Connection | None = conn
-        self.table = table
-
-    def __init_subclass__(cls) -> None:
-        raise MistakeError(
-            "Model cannot be subclassed, did you mean ViewModel?",
-            hint=make_hint("This should be ViewModel, not Model"),
-        )
-
-    def __call__(self, *args: P.args, **kwargs: P.kwargs) -> T:
-        return self.obj(*args, **kwargs)  # type: ignore
-
-    async def find(
-        self,
-        limit: int = -1,
-        offset: int | None = None,
-        *args: P.args,
-        **kwargs: P.kwargs,
-    ) -> list[T]:
-        ...
-
-
-class _Driver(ABC):
-    @classmethod
-    @abstractmethod
-    def ensure(cls) -> None:
-        ...
-
-    @abstractmethod
-    async def find(
-        self, where: Where, limit: int = -1, offset: int = 0
-    ) -> list[dict[str, Any]]:
-        ...
-
-    @abstractmethod
-    async def create(self, where: Where, values: Value) -> None:
-        ...
-
-    @abstractmethod
-    async def update(self, where: Where, values: Value) -> None:
-        ...
-
-    @abstractmethod
-    async def delete(self, where: Where) -> None:
-        ...
-
-
-class MongoDriver(_Driver):
-    pymongo: ClassVar[Any]
-
-    @classmethod
-    def ensure(cls):
-        cls.pymongo = attempt_import("pymongo")
-
-    async def create(self, where: Where, values: Value) -> None:
-        ...
+__all__ = ("Model",)
 
 
 class _Connection(ABC):
     @abstractmethod
-    def __init__(self) -> None:
+    def connect(self) -> None:
         ...
 
     @abstractmethod
@@ -108,132 +27,13 @@ class _Connection(ABC):
         ...
 
 
-def _init(self: Any, *args: Any, **kwargs: Any):
-    ...
-
-
-class _TableNameTransport(Generic[P, T]):
-    def __init__(self, name: str, ob: Model[P, T] | Callable[P, T]):
-        self.name = name
-        self.ob = ob
-
-
-def table(
-    name: str,
-):
-    @overload
-    def wrapper(ob: Model[P, T]) -> Model[P, T]:
-        ...
-
-    @overload
-    def wrapper(ob: Callable[P, T]) -> _TableNameTransport[P, T]:
-        ...
-
-    def wrapper(
-        ob: Model[P, T] | Callable[P, T]
-    ) -> Model[P, T] | _TableNameTransport[P, T]:
-        if isinstance(ob, Model):
-            ob.table = name
-            return ob
-        return _TableNameTransport(name, ob)
-
-    return wrapper
-
-
-@dataclass_transform()
-def _transform(cls: type[A]) -> type[A]:
-    cls.__init__ = _init
-    return cls
-
-
-@overload
-def model(
-    ob_or_none: Callable[P, T] | _TableNameTransport,
-    *,
-    conn: _Connection | None = None,
-    table_name: str = "view",
-) -> Model[P, T]:
-    ...
-
-
-@overload
-def model(
-    ob_or_none: None = ...,
-    *,
-    conn: _Connection | None = None,
-    table_name: str = "view",
-) -> Callable[[Callable[P, T]], Model[P, T]]:
-    ...
-
-
-def model(
-    ob_or_none: Callable[P, T] | None | _TableNameTransport = None,
-    *,
-    conn: _Connection | None = None,
-    table_name: str = "view",
-) -> Model[P, T] | Callable[[Callable[P, T]], Model[P, T]]:
-    def impl(raw_ob: Callable[P, T] | _TableNameTransport) -> Model[P, T]:
-        if isinstance(raw_ob, _TableNameTransport):
-            name = raw_ob.name
-            ob = raw_ob.ob
-        else:
-            name = table_name
-            ob = raw_ob
-
-        if not issubclass(cast(type, ob), ViewModel):
-            if isinstance(ob, type):
-                msg = f"{ob.__name__} does not inherit from ViewModel, did you forget?"  # noqa
-
-            msg = f"{ob!r} does not inherit from ViewModel, did you forget?"
-
-            raise MistakeError(
-                msg, hint=make_hint("This should inherit from ViewModel")
-            )
-
-        body = get_body(ob)  # type: ignore
-        ob = _transform(ob)  # type: ignore
-
-        for k, v in body.items():
-            ...
-
-        return Model(ob, name, conn=conn)
-
-    if ob_or_none: 
-        return impl(ob_or_none)
-
-    return impl
-
-
-
-#database
-import psycopg2  # Import the PostgreSQL driver library
-import sqlite3  # Import the SQLite3 driver library
-import mysql.connector # Import the MySQL driver library
-import pymongo # Import the MongoDB driver library
-import asyncio
-
-
-class PostgresConnection(_Connection):
-    '''
-    This class is used to connect to a PostgreSQL database. Pass the connection details as parameters to the constructor. \n
-        ``` 
-        obj = PostgresConnection(
-            "database": **db_name**,
-            "user": **user_name**,
-            "password": **password**,
-            "host": **host_name**,
-            "port": **port_num**
-        )
-        
-        ```
-    '''
-
+class _PostgresConnection(_Connection):
     def __init__(
-        self, 
-        database: str  | None = None,
-        user: str  | None = None,
-        password: str  | None = None,
-        host: str  | None = None,
+        self,
+        database: str | None = None,
+        user: str | None = None,
+        password: str | None = None,
+        host: str | None = None,
         port: int | None = None,
     ) -> None:
         self.database = database
@@ -245,7 +45,6 @@ class PostgresConnection(_Connection):
         self.cursor = None
 
     def create_database_connection(self):
-        # ASGI error comes if I put this in connect method
         return psycopg2.connect(
             database=self.database,
             user=self.user,
@@ -254,78 +53,53 @@ class PostgresConnection(_Connection):
             port=self.port,
         )
 
-    async def connect(self):
-        '''
-            This method is used to connect to the database. \n
-        '''
+    async def connect(self) -> None:
         try:
-            self.connection = await asyncio.to_thread(self.create_database_connection)
-            self.cursor = await asyncio.to_thread(self.connection.cursor)
+            self.connection = await asyncio.to_thread(
+                self.create_database_connection
+            )
+            self.cursor = await asyncio.to_thread(self.connection.cursor)  # type: ignore
         except psycopg2.Error as e:
-            raise ValueError("Unable to connect to the database - invalid credentials") from e
-        
-    
-    async def close(self):
-        '''
-            This method is used to close the connection to the database. \n
-        '''
+            raise ValueError(
+                "Unable to connect to the database - invalid credentials"
+            ) from e
+
+    async def close(self) -> None:
         if self.connection is not None:
             await asyncio.to_thread(self.connection.close)
             self.connection = None
             self.cursor = None
 
 
-
-
-class SQLiteConnection:
-    '''
-    This class is used to connect to an SQLite database. Pass the database file path as a parameter to the constructor.
-    Example usage:
-    
-    obj = SQLiteConnection("mydb.sqlite")
-    '''
-    
-    def __init__(self, database_file: str):
+class _SQLiteConnection:
+    def __init__(self, database_file: str) -> None:
         self.database_file = database_file
-        self.connection = None
-        self.cursor = None
+        self.connection: sqlite3.Connection | None = None
+        self.cursor: sqlite3.Cursor | None = None
 
-    async def connect(self):
-        '''
-        This method is used to connect to the SQLite database.
-        '''
-        try:
-            self.connection = await asyncio.to_thread(sqlite3.connect, self.database_file)
-            self.cursor = await asyncio.to_thread(self.connection.cursor)
-        except sqlite3.Error as e:
-            raise ValueError("Unable to connect to the database - invalid file or permissions") from e
+    async def connect(self) -> None:
+        self.connection = await asyncio.to_thread(
+            sqlite3.connect, self.database_file
+        )
+        self.cursor = await asyncio.to_thread(self.connection.cursor)  # type: ignore
 
-    async def close(self):
-        '''
-        This method is used to close the connection to the database.
-        '''
+    async def close(self) -> None:
         if self.connection is not None:
+            assert self.cursor is not None
             await asyncio.to_thread(self.cursor.close)
             await asyncio.to_thread(self.connection.close)
             self.connection = None
             self.cursor = None
 
 
-
-class MySQLConnection:
-    '''
-    This class is used to connect to a MySQL database. Pass the connection details as parameters to the constructor.
-    Example usage:
-    
-    obj = MySQLConnection(
-        host="localhost",
-        user="username",
-        password="password",
-        database="mydb"
-    )
-    '''
-    
-    def __init__(self, host: str, user: str, password: str, database: str):
+class _MySQLConnection:
+    def __init__(
+        self,
+        host: str,
+        user: str,
+        password: str,
+        database: str,
+    ) -> None:
         self.host = host
         self.user = user
         self.password = password
@@ -333,50 +107,40 @@ class MySQLConnection:
         self.connection = None
         self.cursor = None
 
-    async def connect(self):
-        '''
-        This method is used to connect to the MySQL database.
-        '''
+    async def connect(self) -> None:
         try:
             self.connection = await asyncio.to_thread(
                 mysql.connector.connect,
                 host=self.host,
                 user=self.user,
                 password=self.password,
-                database=self.database
+                database=self.database,
             )
+
             self.cursor = await asyncio.to_thread(self.connection.cursor)
         except mysql.connector.Error as e:
-            raise ValueError("Unable to connect to the database - invalid credentials") from e
+            raise ValueError(
+                "Unable to connect to the database - invalid credentials"
+            ) from e
 
     async def close(self):
-        '''
-        This method is used to close the connection to the database.
-        '''
         if self.connection is not None:
+            assert self.cursor is not None
             await asyncio.to_thread(self.cursor.close)
             await asyncio.to_thread(self.connection.close)
             self.connection = None
             self.cursor = None
 
 
-
-class MongoDBConnection:
-    '''
-    This class is used to connect to a MongoDB database. Pass the connection details as parameters to the constructor.
-    Example usage:
-    ```
-    obj = MongoDBConnection(
-        host="localhost",
-        port=27017,
-        username="username",
-        password="password",
-        database="mydb"
-    )
-    ```
-    '''
-
-    def __init__(self, host: str, port: int, username: str, password: str, database: str):
+class _MongoDBConnection:
+    def __init__(
+        self,
+        host: str,
+        port: int,
+        username: str,
+        password: str,
+        database: str,
+    ):
         self.host = host
         self.port = port
         self.username = username
@@ -386,28 +150,79 @@ class MongoDBConnection:
         self.db = None
 
     async def connect(self):
-        '''
-        This method is used to connect to the MongoDB database.
-        '''
-        try:
-            self.client = await asyncio.to_thread(
-                pymongo.MongoClient,
-                host=self.host,
-                port=self.port,
-                username=self.username,
-                password=self.password,
-                authSource=self.database
-            )
-            self.db = self.client[self.database]
-            print("connection mongo done")
-        except pymongo.errors.ConnectionFailure as e:
-            raise ValueError("Unable to connect to the database - connection failure") from e
+        self.client = await asyncio.to_thread(
+            pymongo.MongoClient,
+            host=self.host,
+            port=self.port,
+            username=self.username,
+            password=self.password,
+            authSource=self.database,
+        )
+        self.db = self.client[self.database]
 
     async def close(self):
-        '''
-        This method is used to close the connection to the MongoDB database.
-        '''
         if self.client is not None:
             await asyncio.to_thread(self.client.close)
             self.client = None
             self.db = None
+
+
+class _Meta(Enum):
+    HASH = 0
+    ID = 1
+
+
+class _ModelMeta:
+    def __init__(self, tp: _Meta):
+        self.tp = tp
+
+
+T = TypeVar("T")
+Hashed = Annotated[T, _ModelMeta(_Meta.HASH)]
+Id = Annotated[T, _ModelMeta(_Meta.ID)]
+
+
+@dataclass_transform()
+class Model:
+    view_initialized: ClassVar[bool] = False
+    __view_body__: ClassVar[ViewBody] = {}
+    view_conn: Union[_Connection, None] = None
+
+    def __init__(self, *args: Any, **kwargs: Any):
+        ...
+
+    def __init_subclass__(cls):
+        model_hints = get_type_hints(Model)
+        actual_hints = get_type_hints(cls)
+        params = {
+            k: actual_hints[k]
+            for k in (model_hints.keys() ^ actual_hints.keys())
+        }
+
+        for k, v in params.items():
+            df = cls.__dict__.get(k)
+            if df:
+                cls.__view_body__[k] = BodyParam(types=v, default=df)
+            else:
+                cls.__view_body__[k] = v
+
+    @classmethod
+    def find(cls) -> list[Self]:
+        ...
+
+    @classmethod
+    def unique(cls) -> Self:
+        ...
+
+    def exists(self) -> bool:
+        ...
+
+    def save(self) -> None:
+        ...
+
+    def json(self) -> dict[str, Any]:
+        ...
+
+    @classmethod
+    def from_json(cls, json: dict[str, Any]) -> Self:
+        ...